refactoring, cleaning and checking
