+++ AskQuestion.php
@@ -1,12 +1,8 @@
 <?php
     if (isset($_GET['id'])){
+        require 'ConnectDatabase.php';
         $update = true;
         $id = $_GET['id'];
-        $username = 'root';
-        $password = '';
-        $database = 'stackex';
-        mysql_connect('localhost', $username, $password);
-        @mysql_select_db($database) or die("Unable to Select Database");
         $query = "SELECT * FROM questions WHERE ID = $id";
         $result = mysql_query($query);
         $topic = mysql_result($result, 0,"topic");
@@ -28,38 +24,8 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <script type = "text/javascript">
-        function validateEmail(email) {
-            var re = /^([\w-]+(?:\.[\w-]+)*)@((?:[\w-]+\.)*\w[\w-]{0,66})\.([a-z]{2,6}(?:\.[a-z]{2})?)$/i;
-            return re.test(email);
-        }
-        function validateForm(){
-            var x = document.forms["q_form"]["name"].value;
-            var y = document.forms["q_form"]["email"].value;
-            var z = document.forms["q_form"]["topic"].value;
-            var a = document.forms["q_form"]["content"].value;
-            if(x==null || x==''){
-                alert("Name must be filled");
-                return false;
-            }
-            if(y==null || y==''){
-                alert("Email must be filled");
-                return false;
-            }else if(!validateEmail(y)){
-                alert("Email is not a valid email");
-                return false;
-            }
-            if(z==null || z==''){
-                alert("Topic must be filled");
-                return false;
-            }
-            if(a==null || a==''){
-                alert("Content must be filled");
-                return false;
-            }
-        }
-    </script>
-    <link rel ="stylesheet" type="text/css" href="Index.css">
+    <script type = "text/javascript" src="Assets/Validator.js"></script>
+    <link rel ="stylesheet" type="text/css" href="Assets/Index.css">
     <title>Simple StackExchange</title>
 
 </head>
@@ -67,11 +33,11 @@
 <body>
 <div id = "container">
     <div id  = "header">
-        <span id  = "logo"> Simple StackExchange </span>
+        <a id  = "logo" href = "index.php"> Simple StackExchange </a>
     </div>
     <div id = "content">
         <h2>What's Your Question?</h2>
-        <form name = "q_form" action="adding.php" onsubmit="return validateForm()" method = "post">
+        <form name = "q_form" action="Data_Manipulation/adding_question.php" onsubmit="return validate_QForm()" method = "post">
             <input type = "hidden" name = "id" value = "<?php echo $id; ?>">
             <input type = "hidden" name = "update" value = "<?php echo $update; ?>">
             <input type = "text" name = "name" placeholder = "Name" value = "<?php echo $author; ?>">
